+++ classes/components/adapter/BaseAdapter.class.php
@@ -3,6 +3,7 @@ namespace Flex\Components\Adapter;
 
 use Flex\Annona\Paging\Relation;
 class BaseAdapter {
+    public const __version = '0.1';
     public array $relation = [];
 
     public function relation(int $total_record, int $page, int $limit, int $block) : Relation{