+++ test/db_where_helper.php
@@ -21,32 +21,32 @@ Log::options([
 ]);
 
 # db where 구문 만들기
-// $whereHelper = new \Flex\Annona\Db\WhereHelper();
-// $where = $whereHelper
-// ->begin('AND')
-//     ->case('name','IN','홍길동,유관순')
-//     ->case('price','>','0')
-//     ->case('signdate','is not','NULL')->end()
-// ->begin('OR')
-//     ->case('price','IN',[1,2,3,4,5,6])
-//     ->case('price_month','>=',7)->end()
-// ->begin('OR')
-//     ->case('title','LIKE',['이순신','대통령'])->end()
-// // ->fetch();
-// ->where;
-// Log::d( $where );
+$whereHelper = new \Flex\Annona\Db\WhereHelper();
+$where = $whereHelper
+->begin('AND')
+    ->case('name','IN','홍길동,유관순')
+    ->case('price','>','0')
+    ->case('signdate','is not','NULL')->end()
+->begin('OR')
+    ->case('price','IN',[1,2,3,4,5,6])
+    ->case('price_month','>=',7)->end()
+->begin('OR')
+    ->case('title','LIKE',['이순신','대통령'])->end()
+// ->fetch();
+->where;
+Log::d( $where );
 
 // # direct
-// $where2 = (new \Flex\Annona\Db\WhereHelper())
-// ->begin('AND')
-//     ->case('price','>','0')->end()
-// ->begin('OR')
-//     ->case('price_month','>=',7)->end()
-// ->begin('OR')
-//     ->case('title','LIKE',['이순신','대통령'])->end()
-// // ->fetch();
-// ->where;
-// Log::d( $where2 );
+$where2 = (new \Flex\Annona\Db\WhereHelper())
+->begin('AND')
+    ->case('price','>','0')->end()
+->begin('OR')
+    ->case('price_month','>=',7)->end()
+->begin('OR')
+    ->case('title','LIKE',['이순신','대통령'])->end()
+// ->fetch();
+->where;
+Log::d( $where2 );
 
 $whereHelper = new WhereHelper();
 $whereHelper->begin('AND')->case('userid', '=', 'test@ddd.com')->end();