+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [1.3]
 
+### - 2024-08-29
+- WhereHelper 클래스 업데이트 WhereHelperInterface 를 통해 확장 가능성 높임
+- DbBaseAdapter 클래스에 WhereHelper 클래스 자동 선언 및 WhereHelper를 상속받은 CustumWhereHelper 클래스 등록 할 수 있도록 기능 추가
+- WhereHelper v2.0 업데이트
+
 ### - 2024-08-28
 - 디비데이터에 따른 타입 지정 샘플 코드 추가
 - Adapter 개념 도입